services/drift/requirements.txt
mlflow==2.14.3
pandas==2.2.2
psycopg2-binary==2.9.9
prometheus-client==0.20.0
